+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: schaya <schaya@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/01 17:45:46 by schaya            #+#    #+#             */
-/*   Updated: 2020/11/20 16:13:38 by schaya           ###   ########.fr       */
+/*   Updated: 2020/11/20 19:27:41 by schaya           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *memptr, int val, size_t num)
 {
-	int					i;
+	size_t				i;
 	unsigned char		*dest;
 	unsigned char		y;
